- added more animals woot woot

+++ userbot/plugins/memes/animals.py
@@ -1,9 +1,9 @@
 import asyncio
 import re
-from time import sleep
 
 from pyrogram import filters
 from pyrogram.types import Message
+
 from userbot import UserBot
 from userbot.helpers.PyroHelpers import ReplyCheck
 from userbot.helpers.aiohttp_helper import AioHttp
@@ -12,41 +12,38 @@ from userbot.plugins.help import add_command_help
 animal = r"([^.]*)$"
 ok_exts = ["jpg", "jpeg", "png"]
 
-animals_data = {
-    "dog": {"url": "https://random.dog/woof.json", "key": "url"},
-    "cat": {"url": "http://aws.random.cat/meow", "key": "file"},
-    "panda": {"url": "https://some-random-api.ml/img/panda", "key": "link"},
-    "redpanda": {"url": "https://some-random-api.ml/img/red_panda", "key": "link"},
-    "bird": {"url": "https://some-random-api.ml/img/birb", "key": "link"},
-    "fox": {"url": "https://some-random-api.ml/img/fox", "key": "link"},
-    "koala": {"url": "https://some-random-api.ml/img/koala", "key": "link"},
-}
+animals_without_facts = ['dog', 'cat', 'panda', 'fox', 'red_panda', 'birb', 'koala', 'kangaroo', 'racoon']
 
-animals = [x for x in animals_data]
+animals_with_facts = ['dog', 'cat', 'panda', 'fox', 'birb', 'koala', 'kangaroo', 'racoon', 'elephant', 'giraffe',
+                      'whale']
 
 
-async def prep_animal_image(animal_data):
+async def prep_animal_image(input_animal):
     ext = ""
     image = None
+
     while ext not in ok_exts:
-        data = await AioHttp().get_json(animal_data["url"])
-        image = data[animal_data["key"]]
+        data = await AioHttp().get_json(f"https://some-random-api.ml/animal/{input_animal}")
+        image = data['image']
         ext = re.search(animal, image).group(1).lower()
+
     return image
 
 
-@UserBot.on_message(filters.command(animals, [".", ""]) & filters.me)
+@UserBot.on_message(filters.command(animals_without_facts, [".", ""]) & filters.me)
 async def animal_image(_, message: Message):
+    cmd = message.command
+
     if len(message.command) > 1:
         return
 
-    animal_data = animals_data[message.command[0]]
-    await message.delete()
-    await UserBot.send_photo(
-        chat_id=message.chat.id,
-        photo=await prep_animal_image(animal_data),
-        reply_to_message_id=ReplyCheck(message),
-    )
+    if cmd[0].lower() in animals_without_facts:
+        await message.delete()
+        await UserBot.send_photo(
+            chat_id=message.chat.id,
+            photo=await prep_animal_image(cmd[0].lower()),
+            reply_to_message_id=ReplyCheck(message),
+        )
 
 
 @UserBot.on_message(filters.command("fact", ".") & filters.me)
@@ -59,17 +56,19 @@ async def fact(_, message: Message):
         await message.delete()
         return
 
-    await message.edit(f"```Getting {cmd[1]} fact```")
-    link = "https://some-random-api.ml/facts/{animal}"
+    if cmd[1].lower() in animals_with_facts:
+        await message.edit(f"```Getting {cmd[1]} fact```")
 
-    if cmd[1].lower() in animals:
+        link = "https://some-random-api.ml/facts/{animal}"
+
         fact_link = link.format(animal=cmd[1].lower())
+
         try:
             data = await AioHttp().get_json(fact_link)
             fact_text = data["fact"]
         except Exception:
             await message.edit("```The fact API could not be reached```")
-            sleep(3)
+            await asyncio.sleep(3)
             await message.delete()
         else:
             await message.edit(fact_text, disable_web_page_preview=True)
@@ -81,7 +80,7 @@ async def fact(_, message: Message):
 
 # Animal image help
 animal_image_help = []
-for x in animals:
+for x in animals_without_facts:
     animal_image_help.append([f".{x}", f"Sends a random picture of a {x}"])
 
 animal_image_help.append(["These commands", "Works without the command prefix also"])
@@ -92,7 +91,6 @@ add_command_help(
     "facts",
     [
         [f".fact {x}", f"Send a random fact about {x}"]
-        for x in animals
-        if x != "redpanda"
+        for x in animals_with_facts
     ],
 )
